+++ CHANGELOG.md
@@ -52,6 +52,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - allow passing multiple calldata argument items without quotes
 - display correct error message when account file is invalid
 
+#### Changed
+- `start_warp` and `stop_warp` now take `CheatTarget` as the first argument instead of `ContractAddress`. Read more [here](./docs/src/appendix/cheatcodes/start_warp.md). 
+
+
 ## [0.9.1] - 2023-10-30
 
 ### Forge